Fix null warning in studio
CreateInstance obviously can't know if the constructor of an arbitrary class returns null. So obviously the return value is nullable. And in theory we could have made a mistake in our class, so a null value is produced. Hence adding a "is not null" check here to silence the studio null warning is correct.  Even more correct would be to raise or log some kind of error if null happened, but I haven't looked into the error raising conventions of the API.